pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "company-researcher"
version = "1.0.0"
description = "Multi-agent AI system for comprehensive company research"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Company Researcher Team"}
]
keywords = [
    "ai",
    "agents",
    "research",
    "llm",
    "langchain",
    "langgraph",
    "company-research",
    "multi-agent"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "anthropic>=0.18.0",
    "langchain>=0.1.0",
    "langchain-anthropic>=0.1.0",
    "langchain-community>=0.0.20",
    "langgraph>=0.0.20",
    "tavily-python>=0.3.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "black>=23.0.0",
    "pre-commit>=3.0.0",
]
api = [
    "fastapi>=0.100.0",
    "uvicorn>=0.23.0",
    "websockets>=11.0",
]
output = [
    "reportlab>=4.0.0",
    "openpyxl>=3.1.0",
    "python-pptx>=0.6.21",
]
memory = [
    "chromadb>=0.4.0",
    "openai>=1.0.0",  # For embeddings
]
monitoring = [
    "agentops>=0.1.0",
    "langsmith>=0.1.0",
    "psutil>=5.9.0",
]
all = [
    "company-researcher[dev,api,output,memory,monitoring]",
]

[project.urls]
Homepage = "https://github.com/company-researcher/company-researcher"
Documentation = "https://company-researcher.readthedocs.io"
Repository = "https://github.com/company-researcher/company-researcher"
Issues = "https://github.com/company-researcher/company-researcher/issues"

[project.scripts]
company-researcher = "company_researcher.api.app:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

# ============================================================================
# Pytest Configuration
# ============================================================================

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--tb=short",
    "--strict-markers",
    "-ra",
]
markers = [
    "unit: Unit tests (fast, no external dependencies)",
    "integration: Integration tests (may require external services)",
    "slow: Slow tests (skip with -m 'not slow')",
    "requires_api: Tests requiring external API keys",
    "requires_llm: Tests requiring LLM access",
]
filterwarnings = [
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
]
asyncio_mode = "auto"

# ============================================================================
# Coverage Configuration
# ============================================================================

[tool.coverage.run]
source = ["src/company_researcher"]
branch = true
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
show_missing = true

# ============================================================================
# Ruff Configuration (Linting)
# ============================================================================

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG",    # flake8-unused-arguments
    "SIM",    # flake8-simplify
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "B904",   # raise without from inside except
    "ARG001", # unused function argument
]

[tool.ruff.per-file-ignores]
"tests/*" = ["ARG", "S101"]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["company_researcher"]

# ============================================================================
# Black Configuration (Formatting)
# ============================================================================

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]
include = '\.pyi?$'
exclude = '''
/(
    \.git
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | venv
    | _build
    | buck-out
    | build
    | dist
    | __pycache__
)/
'''

# ============================================================================
# MyPy Configuration (Type Checking)
# ============================================================================

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
exclude = [
    "venv",
    ".venv",
    "__pycache__",
    ".git",
    ".archive",
]

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
